+++ src/Constants/Formats.php
@@ -2,11 +2,24 @@
 
 namespace Sysbot\Bin\Constants;
 
+/**
+ * Class Formats
+ * @package Sysbot\Bin\Constants
+ */
 class Formats
 {
 
+    /**
+     * Identifies a 32-bit integer.
+     */
     public const LONG = 'long';
+    /**
+     * Identifies a 64-bit integer.
+     */
     public const LONG_LONG = 'long_long';
+    /**
+     * Identifies a string.
+     */
     public const STRING = 'string';
 
 }